requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-JWT-Extended==4.2.1
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
PyJWT==2.1.0
pymongo==3.11.4
python-bsonjs==0.2.0
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
